+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update `benchmarks/`
 - Fix inherited properties override in `RequestClient`
 - Fix catching of errors in `RestResponse.deserializeBody`
+- Fix `JsonModelSerializer.merge` where this deserializers from this where merge was invoked where ignored
 
 ## 0.3.2+1